3-feature-pipeline/pyproject.toml
[tool.poetry]
name = "3-feature-pipeline"
version = "0.1.0"
description = "Bytewax Feature Processing Pipeline"
authors = [""]

[tool.poetry.dependencies]
# Python Version
python = ">=3.10, <3.12"

# Core Data Processing
numpy = "^1.26.4"
pandas = "^2.2.2"
pydantic = "^2.6.3"
pydantic-settings = "^2.1.0"

# Streaming Processing
bytewax = "0.18.2"

# Database Connections
pymongo = "^4.7.1"
pika = "^1.3.2"
qdrant-client = "^1.8.0"

# Machine Learning & Embeddings
sentence-transformers = "^2.6.1"
open-clip-torch = "^2.24.0"

# LangChain dependencies
langchain = "^0.1.13"
langchain-text-splitters = "^0.0.2"
langchain-community = "^0.0.36"

# Document Processing
unstructured = "^0.12.6"

# Image Processing
Pillow = "^10.2.0"
opencv-python-headless = "^4.9.0"

# Optional Machine Learning
torch = {version = ">=2.0.0, <3.0.0", optional = true}
transformers = {version = "^4.38.0", optional = true}

# Logging & Utilities
structlog = "^24.1.0"
rich = "^13.7.1"

[tool.poetry.extras]
ml = ["torch", "transformers"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
